% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accessions}
\alias{accessions}
\title{Collection data for the 1001 genomes project accessions}
\format{A data frame with 1135 rows and 8 variables
\describe{
  \item{Ecotype.ID}{unique number for each accession which corresponds to
  the "Indiv" column in the VCF}
  \item{Name}{common name of the accession}
  \item{CS.Number}{germplasm identifier for ordering seeds from ABRC or
  NASC}
  \item{Country}{country in which the accession was initially collected}
  \item{Lat}{latitude of initial collection}
  \item{Long}{longitude of initial collection}
  \item{Collector}{name of collector}
  \item{Sequenced.by}{institution which sequenced the accession}
}}
\source{
<http://1001genomes.org/accessions.html>
}
\usage{
accessions
}
\description{
A dataset containing unique identifiers, names, locations, and collectors information for each of the 1135 accessions in the 1001 genomes project.
}
\keyword{datasets}
