% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{addAlnPosToAnno}
\alias{addAlnPosToAnno}
\title{Add alignment positions to an annotation data frame}
\usage{
addAlnPosToAnno(anno_df, aln_df)
}
\arguments{
\item{aln_df}{}
}
\description{
Add alignment positions to an annotation data frame
}
