% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{alignCDS}
\alias{alignCDS}
\title{Make an alignment of Coding sequences}
\usage{
alignCDS(IDs)
}
\arguments{
\item{IDs}{a vector of TAIR transcript IDs, e.g. "AT3G26890.1"}
}
\value{
aligned CDS and amino acid sequences as a list of XStringSet objects
}
\description{
Make an alignment of Coding sequences
}
\examples{
IDs <- c("AT3G62980.1", "AT3G26810.1")
alignment <- alignCDS(IDs)
browseSeqs(alignment[[2]])
}
