% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{approxRefGt}
\alias{approxRefGt}
\title{return a count of accessions with undetermined genotype or 0 depending on if there are 0|0 genotpyes present}
\usage{
approxRefGt(groupFreq)
}
\arguments{
\item{groupFreq}{a group GT_Frequencies tibble for a single position}
}
\value{
0 if 0|0 genotypes are present or (1135 - number of determined genotypes) if no 0|0 genotypes are present.
}
\description{
return a count of accessions with undetermined genotype or 0 depending on if there are 0|0 genotpyes present
}
