% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{calcDiversity}
\alias{calcDiversity}
\title{Calculate Nei's nucleotide diversity statistic for a single position,
given a vector of counts of uninque genotypes at that location}
\usage{
calcDiversity(GTfreq)
}
\arguments{
\item{GTfreq}{numeric vector listing counts of uninque genotypes}
}
\value{
numeric value nucleotide diversity
}
\description{
Calculate Nei's nucleotide diversity statistic for a single position,
given a vector of counts of uninque genotypes at that location
}
