% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{chunkAlnDF}
\alias{chunkAlnDF}
\title{Chunk up an alignment data frame to allow facetting over
multiple rows}
\usage{
chunkAlnDF(aln_df, chunk_width = 80)
}
\arguments{
\item{aln_df}{An alignment data frame}

\item{chunk_width}{The number of units (bp or aa) in a chunk.}
}
\value{
\code{aln_df} with the addition of a \code{chunk} column for
 wrapping the alignment across facets
}
\description{
Chunk up an alignment data frame to allow facetting over
multiple rows
}
\examples{
#' # make an alignment
IDs <- c("AT3G62980.1", "AT3G26810.1")
alignment <- alignCDS(IDs)
# make an alignment data frame
aln_df <- makeAlnDF(alignment[[2]])
chunkAlnDF(aln_df)

}
