% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{chunkAnnotation}
\alias{chunkAnnotation}
\title{Add a chunks column to an annotation data frame}
\usage{
chunkAnnotation(anno_df, chunks)
}
\arguments{
\item{anno_df}{an annotation data frame from \link{readAnnotationFile}
with attached alignment positions from \link{addAlnPosToAnno}}

\item{chunks}{a data frame of chunks to facet an alignment/annotation
data frame across}
}
\value{
anno_df with the addition of chunks columns (and potentially extra
rows) for facetting
}
\description{
This function adds chunk information for wrapping an alignment and
annotations across multiple rows. This will also split individual
annotations across rows.
}
\examples{
## first make a gene info DF
geneInfo <- getGeneInfo(genes = c("AT3G62980", "AT3G26810"))

## read in an annotation DF
anno_df <- readAnnotationFile(filename = system.file("extdata",
"AFB_annotations.csv", package = "r1001genomes"), gene_info = geneInfo )

## make an alignment DF
IDs <- c("AT3G62980.1", "AT3G26810.1")
alignment <- alignCDS(IDs)
aln_df <- makeAlnDF(alignment[[2]])
aln_df <- chunkAlnDF(aln_df)
chunks <- makeChunksDF(aln_df)

anno_df <- addAlnPosToAnno(anno_df, aln_df)

chunkAnnotation(anno_df, chunks)

}
