% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{coding_Diversity_Plot}
\alias{coding_Diversity_Plot}
\title{Title}
\usage{
coding_Diversity_Plot(data)
}
\arguments{
\item{data}{tidy tibble/df with EFF field parsed and diversity calculated,
e.g.
\code{myvcf <- VCFByTranscript(geneInfo[1, ], strains)
 mydata <- parseEFF(myvcf$dat, geneInfo$transcript_ID)
 mydata <- Nucleotide_diversity(mydata)}}
}
\description{
Title
}
