% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{codonNumberKernel}
\alias{codonNumberKernel}
\title{Add the codon number to a dataframe containing an "Amino_Acid_Change" column}
\usage{
codonNumberKernel(SNPData)
}
\arguments{
\item{SNPData}{}
}
\description{
Add the codon number to a dataframe containing an "Amino_Acid_Change" column
}
