% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{diversity_calc}
\alias{diversity_calc}
\title{Calculate nucleotide diversity}
\usage{
diversity_calc(GT_freq)
}
\arguments{
\item{GT_freq}{the vector of genotype frequencies at a specific site}
}
\value{
the nucleotide diversity at that site
}
\description{
Calculate nucleotide diversity
}
