% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{downloadData}
\alias{downloadData}
\title{Download and save a .VCF file from 1001genomes.org.}
\usage{
downloadData(fName, strainStr, regionStr, download = TRUE)
}
\arguments{
\item{fName}{the file name the downloaded vcf will be saved as eg.
"data1.vcf"}

\item{strainStr}{A character string of comma separated strain or
"Ecotype ID"s used in the construction of the URL}

\item{regionStr}{A character string of the format "[chrom]:[start]-[stop]"
where [chrom], [start], and [stop] are numbers. this string is used
directly in the construction of the URL}

\item{download}{A logical, if TRUE (default) the file will be downloaded, if
FALSE, the file will not be downloaded, the URL will still be returned}
}
\value{
The URL used for the download
}
\description{
Download and save a .VCF file from 1001genomes.org.
}
