% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{downloadMerge}
\alias{downloadMerge}
\title{Download two vcf files from 1001genomes.org and merge their contents}
\usage{
downloadMerge(fName, strainVect, regionStr)
}
\arguments{
\item{fName}{a character string of the file name to save the final .vcf file to}

\item{strainVect}{numeric vector list of the strains}

\item{regionStr}{A character string of the format "[chrom]:[start]-[stop]"
where [chrom], [start], and [stop] are numbers. this string is used
directly in the construction of the URL}
}
\value{
a vcfR object of teh combined vcf file
}
\description{
Download two vcf files from 1001genomes.org and merge their contents
}
\examples{
\\code{my.VCF <- downloadMerge("myFullVCF.vcf.gz", strains, "1:4368760-4371298")}
}
