% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{geneInfoFromFile}
\alias{geneInfoFromFile}
\title{make geneInfo dataframe based on .csv file of tair loci and names/symbols}
\usage{
geneInfoFromFile(fname, firstOnly = TRUE, useCache = TRUE,
  source = "tair10")
}
\arguments{
\item{fname}{filename of csv files containing "tair_locus" and "name" fields}

\item{firstOnly}{logical, if true only return transcript IDs containing ".1"}

\item{useCache}{logical, read from and write to a file of cached genes?}
}
\value{
geneInfo dataframe see getGeneInfo
}
\description{
make geneInfo dataframe based on .csv file of tair loci and names/symbols
}
