% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{geom_str_align}
\alias{geom_str_align}
\title{String alignment geom for plotting XStringSet Alignments}
\usage{
geom_str_align(mapping = NULL, data = NULL)
}
\arguments{
\item{data}{an alignment data.frame made from an XStringSet. In the future,
we plan to add pretty defaults allowing XStringSets to be plotted.}

\item{...}{additional parameters passed on to geoms like aes()}
}
\description{
String alignment geom for plotting XStringSet Alignments
}
