% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr}
\alias{gr}
\title{Araport11 genome annotation}
\format{A granges object with 789890 ranges across 5 nuclear chromosomes as
well as the mitochondrial and chloroplastic chromosomes.}
\usage{
gr
}
\description{
Araport11 genome annotation
}
\keyword{datasets}
