% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{makeAlnDF}
\alias{makeAlnDF}
\title{Make an alignment data frame for plotting}
\usage{
makeAlnDF(alignment)
}
\arguments{
\item{alignment}{an AAStringSet or DNAStringSet resulting from an alignment made with DECIPHER.}
}
\value{
a tidy long data frame of the alignment with columns `seq_name`, `letter`, `aln_pos`, and `seq_pos`, where `aln_pos` is the position of the letter in the alignment of all sequences and `seq_pos` is the position of the letter in the native sequence.
}
\description{
Make an alignment data frame for plotting
}
\examples{
IDs <- c("AT3G62980.1", "AT3G26810.1")
alignment <- alignCDS(IDs)
makeAlnDF(alignment[[2]])

}
