% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{makeChunksDF}
\alias{makeChunksDF}
\title{Create a chunks data frame from a chunked alignment data frame}
\usage{
makeChunksDF(aln_df)
}
\arguments{
\item{chunk_width}{}
}
\value{
a chunks data frame containing the start, end, and chunk name for
each chunk of the alignment
}
\description{
To allow similar chunking of annotation layers for facetting
across multiple rows
}
\examples{
# make an alignment
IDs <- c("AT3G62980.1", "AT3G26810.1")
alignment <- alignCDS(IDs)
# make an alignment data frame
aln_df <- makeAlnDF(alignment[[2]])
aln_df <- chunkAlnDF(aln_df)
makeChunksDF(aln_df)

}
