% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{makeRegionString}
\alias{makeRegionString}
\title{Make a string containing the desired genomic ranges
of the format "chrom:start-stop"}
\usage{
makeRegionString(data)
}
\arguments{
\item{data}{a data.frame with columns `chromosome_name`, `transcript_start`, and `transcript_end`}
}
\value{
a vector of strings with the format "chrom:start-stop"
}
\description{
Make a string containing the desired genomic ranges
of the format "chrom:start-stop"
}
\examples{

}
