% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{Nucleotide_diversity}
\alias{Nucleotide_diversity}
\title{Calculate nucleotide diversity for each site in a tidyVCF}
\usage{
Nucleotide_diversity(tidyVCF.dat)
}
\arguments{
\item{tidyVCF.dat}{The $dat field of a tidyVCF object}
}
\value{
The tidyVCF.dat object with an additional column containing the Nucleotide diversity.
}
\description{
Calculate nucleotide diversity for each site in a tidyVCF
}
