% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{parseEFF}
\alias{parseEFF}
\title{Parse the EFF field of the VCF files from 1001genomes.org}
\usage{
parseEFF(tidyVCF, Transcript_ID)
}
\arguments{
\item{tidyVCF}{tidyVCF$dat tibble to be parsed. should have an "EFF" field in
its $dat section}

\item{Transcript_ID}{Character string of the transcript ID}
}
\value{
new tidyVCF object with added columns for the parsed effect fields
}
\description{
Parse the EFF field of the VCF files from 1001genomes.org
}
