% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{parseEFFKernel}
\alias{parseEFFKernel}
\title{Kernel funtion for EFF field parsing. For internal use.}
\usage{
parseEFFKernel(data, Transcript_ID, EFFColNames)
}
\arguments{
\item{EFFColNames}{}
}
\value{

}
\description{
Kernel funtion for EFF field parsing. For internal use.
}
