% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{plotCodingDiv}
\alias{plotCodingDiv}
\title{Plot nucleotide diversity over the coding sequence by codon}
\usage{
plotCodingDiv(uniqueCodingVars)
}
\arguments{
\item{uniqueCodingVars}{}

\item{uniqueCodingVars}{}
}
\value{
plot object
}
\description{
Plot nucleotide diversity over the coding sequence by codon
}
\examples{
New Plot Coding Div

}
