% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{polymorphRow}
\alias{polymorphRow}
\title{create a row of statistics for a given tidyVCF dataframe including non unique
variant counts and nucleotide diversity values.}
\usage{
polymorphRow(data, geneInfo = NULL)
}
\arguments{
\item{data}{tidyVCF}

\item{geneInfo}{geneInfo dataframe containing a row for the provided tidyVCF gene}
}
\description{
create a row of statistics for a given tidyVCF dataframe including non unique
variant counts and nucleotide diversity values.
}
