% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{promoterVariantToString}
\alias{promoterVariantToString}
\title{make DNAStrings of sequences for each gene of each accession}
\usage{
promoterVariantToString(SNPs, genome, ranges, files_out = TRUE)
}
\arguments{
\item{SNPs}{a VRanges object containing the SNPs you would like to make
sequences for}

\item{genome}{a BSgenomes object corresponding to the genome used to call
the SNPs and also corresponding to the ranges set in 'ranges'}

\item{ranges}{a named GRanges object containing the ranges of the sequence
elements you would like to extract}
}
\description{
make DNAStrings of sequences for each gene of each accession
}
