% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{readAnnotationFile}
\alias{readAnnotationFile}
\title{Read an annotation file}
\usage{
readAnnotationFile(filename, wide = FALSE, domains = TRUE,
  gene_info = NULL)
}
\arguments{
\item{filename}{filename or path to the annotation csv file to be read.
This should by default be a long format csv containing at least columns
named "gene", "annotation", "annotation_type", "position", and
"position_type". "Gene" should match the names of the tracks you wish to
annotate. "Annotation" is a description of the position and may be the start
or end of annotations spanning multiple consecutive positions
(multiposition annotations) (e.g. domain_start), note that the
underscore separator and "start" and "end" are crucial to the processing of
domain annotations and "start" and "end" should not be used in single
position annotations. "Position_type" should be either
"AA" refering to the amino acid position, or "DNA" refering to the coding
sequence.}

\item{wide}{is the annotation file a wide format data frame with colums for
the start and end of each domain? Defaults to \code{FALSE}}

\item{domains}{create multiposition annotations from annotations containing
"start" and "end"? This will allow shading of the entire domain as opposed
to just marking the start and end positions.}
}
\value{
returns an annotation data frame, or if domains = TRUE a list of
data frames with the first "domains" element containing multiposition
annotation and the second "positions" element containing single position
annotations.
}
\description{
Read an annotation file
}
\examples{
readAnnotationFile(filename = system.file("extdata", "AFB_annotations.csv",
                                        package = "r1001genomes"))
}
