% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{relableTairSymbol}
\alias{relableTairSymbol}
\title{Rename TAIR symbols of geneInfo table based on .csv file}
\usage{
relableTairSymbol(geneInfo, fnames)
}
\arguments{
\item{geneInfo}{geneInfo dataframe, see getGeneInfo() function}

\item{fnames}{vector of filenames of csv files containing "tair_locus" and "name" fields}
}
\value{
geneInfo dataframe, where the tair_symbol of the original geneInfo is replaced
}
\description{
Rename TAIR symbols of geneInfo table based on .csv file
}
