% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{run1001genomes}
\alias{run1001genomes}
\title{Run 1001 genomes browser shiny app}
\usage{
run1001genomes()
}
\value{
Runs the app
}
\description{
Run 1001 genomes browser shiny app
}
\examples{
run1001genomes()
}
