% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SNPeff_order}
\alias{SNPeff_order}
\title{Rank order of effects called by SNPeff}
\format{A data frame with 19 rows and 2 variables
\describe{
  \item{strength}{the strength of the effect on function, where higher
  strengths are more likely to be deleterious}
  \item{effect}{the name of the effect}
}}
\usage{
SNPeff_order
}
\description{
Rank order of effects called by SNPeff
}
\keyword{datasets}
