% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{VCFByTranscript}
\alias{VCFByTranscript}
\title{download VCF, optionally in tidyVCF format}
\usage{
VCFByTranscript(geneInfo, strains, tidy = TRUE, dataOnly = TRUE)
}
\arguments{
\item{geneInfo}{single row of geneInfo dataframe}

\item{strains}{numeric vector of strain/ecotypes to include}

\item{tidy}{logical, if true VCF will be provided in tidyVCF format, see
vcfR documentation}
}
\description{
download VCF, optionally in tidyVCF format
}
