% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accessions}
\alias{accessions}
\title{Collection data for the 1001 genomes project accessions}
\format{A data frame with 1135 rows and 8 variables
\describe{
  \item{Ecotype.ID}{unique number for each accession which corresponds to
  the "Indiv" column in the VCF}
  \item{Name}{common name of the accession}
  \item{CS.Number}{germplasm identifier for ordering seeds from ABRC or
  NASC}
  \item{Country}{country in which the accession was initially collected}
  \item{Lat}{latitude of initial collection}
  \item{Long}{longitude of initial collection}
  \item{Collector}{name of collector}
  \item{Sequenced.by}{institution which sequenced the accession}
}}
\source{
\url<https://www.googleapis.com/fusiontables/v2/query?sql=SELECT%20id%20AS%20tg_ecotypeid,%20name,%20CS_number,%20country,%20latitude,%20longitude,%20collector,%20seq_by%20FROM%201oabIAuVSTkoG3qTAqM2sj-qiurysSWc5JEMq-TWm%20ORDER%20by%20id&key=AIzaSyB1GbyVSIOK12RJbFMkaIJjwhVNG-b8fjc&alt=csv>
}
\usage{
accessions
}
\description{
A dataset containing unique identifiers, names, locations, and collectors information for each of the 1135 accessions in the 1001 genomes project.
}
\keyword{datasets}
