% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{addAccDetails}
\alias{addAccDetails}
\title{Add accession metadata to a dataset containing ecotype IDs}
\usage{
addAccDetails(tidyVCF, allAccs = FALSE)
}
\arguments{
\item{tidyVCF}{Tidy format VCF data.}

\item{allAccs}{logical, if `TRUE` include rows for all accessions even if no
rows in the data exist with that accession.}
}
\value{
dataframe with accession details added
}
\description{
Add accession metadata to a dataset containing ecotype IDs
}
\examples{
## make a gene info DF
geneInfo <- getGeneInfo(genes = c("AT3G62980", "AT3G26810"))

## download a single VCF
myVCF <- VCFByTranscript(geneInfo[1, ])

## add accession details to a single VCF
myVCFWithAccDets <- addAccDetails(myVCF)

## This function can also be applied to a list of VCF files created by the
## `VCFList()` function, see example in the documentation for VCFList()

}
