% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{addAccDetails}
\alias{addAccDetails}
\title{Add accession metadata to a dataset containing ecotype IDs}
\usage{
addAccDetails(data, ecotypeColumn = "Indiv")
}
\arguments{
\item{ecotypeColumn}{}
}
\description{
Add accession metadata to a dataset containing ecotype IDs
}
