% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{codonNumberKernel}
\alias{codonNumberKernel}
\title{Add codon numbering to the nucleotide diversity kernel}
\usage{
codonNumberKernel(SNPData)
}
\arguments{
\item{SNPData}{}
}
\description{
Add codon numbering to the nucleotide diversity kernel
}
