% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{diversityStats}
\alias{diversityStats}
\title{calculate nucleotide diversity statsistics for a gene/transcript
use with ldply() on VCFList objects to create a table.}
\usage{
diversityStats(data, geneInfo = NULL)
}
\arguments{
\item{data}{tidyVCF with 'EFF' field parsed and diversity calculated by position}

\item{geneInfo}{}
}
\value{
a row with nucleotide diversity statistics of the transcript.
}
\description{
calculate nucleotide diversity statsistics for a gene/transcript
use with ldply() on VCFList objects to create a table.
}
