% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{geneInfoFromGff}
\alias{geneInfoFromGff}
\title{get gene information from .gff file}
\usage{
geneInfoFromGff(genes, gffFile)
}
\arguments{
\item{genes}{a character vector of tair IDs of the genes to retrieve}

\item{gffFile}{file path to the gff file or gz compressed gff file.}
}
\value{
gene info in same format as getGeneInfo, but including all transcripts.
}
\description{
get gene information from .gff file
}
\examples{
gffFile <- system.file("extdata",
                       "Araport11_GFF3_genes_transposons.201606.gff.gz",
                       package="r1001genomes")
geneInfo <- geneInfoFromGff(genes = c("AT3G62980", "AT3G26810"), gffFile)
}
