% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{getGeneInfo}
\alias{getGeneInfo}
\title{Get gene information}
\usage{
getGeneInfo(genes, firstOnly = TRUE, inputType = "tair_locus",
  useCache = TRUE)
}
\arguments{
\item{useCache}{}
}
\description{
Get the start and end position of each transcript of a set of
 genes using `bioMart` package to query the TAIR10 database via plants.ensembl.org
}
