% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{getGeneInfo}
\alias{getGeneInfo}
\title{Get gene information}
\usage{
getGeneInfo(genes, firstOnly = TRUE, inputType = "tair_locus",
  useCache = TRUE)
}
\arguments{
\item{genes}{a character vector of tair IDs of the genes to retrieve}

\item{firstOnly}{logical, if true only return transcript IDs containing ".1"}

\item{useCache}{logical, read from and write to a file of cached genes?}
}
\description{
Get the start and end position of each transcript of a set of
 genes using `bioMart` package to query the TAIR10 database via plants.ensembl.org
}
