% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{getGeneInfo}
\alias{getGeneInfo}
\title{Get gene information}
\usage{
getGeneInfo(genes, firstOnly = TRUE, useCache = TRUE,
  source = "tair10", verbose = FALSE)
}
\arguments{
\item{genes}{a character vector of tair IDs of the genes to retrieve}

\item{firstOnly}{logical, if true only return transcript IDs containing ".1"}

\item{useCache}{logical, read from and write to a file of cached genes?}

\item{source}{the source to retrieve gene information from. "tair10" will
use the tair 10 database via biomart (requires internet), "araport11" uses a
gff file of the araport11 data stored locally in  the r1001genomes R package
a file path to a .gff or .gff.gz file may also be used.}

\item{verbose}{TRUE or FALSE, print new genes added to cache in console?}
}
\value{
a table containing fields from the TAIR database on the provided genes
including "transcript_ID" and "regionString" columns required for other fuctions in this code
}
\description{
Get the start and end position of each transcript of a set of
 genes using `bioMart` package to query the TAIR10 database via plants.ensembl.org
}
\examples{
geneInfo <- getGeneInfo(genes = c("AT3G62980", "AT3G26810"))

}
