% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{labelBySNPs}
\alias{labelBySNPs}
\title{Label accessions with the variants they contain}
\usage{
labelBySNPs(data, collapse = TRUE)
}
\arguments{
\item{data}{a data frame of variants with a column named "Indiv"}

\item{collapse}{logicial value, should each accession be a single line?}
}
\value{
a df with a single row per accession, with a new column "SNPs" that
 has a single text string detailing all the variants from data within that
 accession
}
\description{
Label accessions with the variants they contain
}
