% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{Nucleotide_diversity}
\alias{Nucleotide_diversity}
\title{calculates site nucleotide diversity for each site}
\usage{
Nucleotide_diversity(tidyVCF)
}
\arguments{
\item{tidyVCF}{the $dat field of a tidyVCF object}
}
\description{
calculates site nucleotide diversity for each site
}
