% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{parseEFF}
\alias{parseEFF}
\title{Parse the EFF field of the VCF files from 1001genomes.org}
\usage{
parseEFF(tidyVCF)
}
\arguments{
\item{tidyVCF}{tidyVCF$dat tibble to be parsed. should have an 'EFF' field}

\item{Transcript_ID}{Character string of the transcript ID. if NULL, the
function will look for it as the attribute "transcript_ID" of the dataframe}
}
\value{
new tidyVCF object with added columns for the parsed effect fields
}
\description{
Parse the EFF field of the VCF files from 1001genomes.org
}
