% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{variantCounts}
\alias{variantCounts}
\title{counts number of variants in certain affect categories
use with ldply() on VCFList objects to create a table.}
\usage{
variantCounts(data, unique = TRUE)
}
\arguments{
\item{data}{tidyVCF with 'EFF' field parsed}

\item{unique}{logical, if true only unique variants will be counted}
}
\description{
counts number of variants in certain affect categories
use with ldply() on VCFList objects to create a table.
}
\examples{

geneInfo <- getGeneInfo(genes = c("AT3G62980", "AT3G26810"))

## download and compile a list of VCF files
myVCFList <- VCFList(geneInfo)

## tabultate unique variants
ldply(myVCFList, variantCounts, unique=TRUE, .id="transcript_ID")

# tabulate total counts of variants relative to the reference sequence.
ldply(myVCFList, variantCounts, unique=TRUE, .id="transcript_ID")

}
