% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Utils.R
\name{VCFList}
\alias{VCFList}
\title{download and store several VCFs in a list structure, named by their transcirpt ID}
\usage{
VCFList(geneInfo, by = "transcript", tidy = TRUE)
}
\arguments{
\item{by}{}
}
\description{
download and store several VCFs in a list structure, named by their transcirpt ID
}
